refactor: package명 변경 example 제거
