--- server/src/main/java/com/example/foodshare/domain/Category.java
@@ -1,33 +1,0 @@
-package com.example.foodshare.domain;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.sql.Timestamp;
-
-
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
-@Entity
-@Table(name = "categories")
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "category_id")
-    private Long categoryId;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "created_at")
-    private Timestamp createdAt;
-
-    @Column(name = "updated_at")
-    private Timestamp updatedAt;
-
-    // Getters and Setters
-}
