+++ server/src/main/java/com/foodshare/domain/Spoon.java
@@ -1,4 +1,4 @@
-package com.example.foodshare.domain;
+package com.foodshare.domain;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -12,16 +12,12 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @Data
 @Entity
-@Table(name = "trade_reviews")
-public class TradeReview {
+@Table(name = "spoon")
+public class Spoon {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "review_id")
-    private Long reviewId;
-
-    @ManyToOne
-    @JoinColumn(name = "food_id", referencedColumnName = "food_id")
-    private Food food;
+    @Column(name = "spoon_id")
+    private Long spoonId;
 
     @ManyToOne
     @JoinColumn(name = "receiver_id", referencedColumnName = "user_id")
@@ -31,11 +27,8 @@ public class TradeReview {
     @JoinColumn(name = "giver_id", referencedColumnName = "user_id")
     private User giver;
 
-    @Column(name = "rating")
-    private Integer rating;
-
-    @Column(name = "comment")
-    private String comment;
+    @Column(name = "score")
+    private Integer score;
 
     @Column(name = "created_at")
     private Timestamp createdAt;
