+++ server/src/main/java/com/foodshare/domain/TradeReview.java
@@ -1,0 +1,47 @@
+package com.foodshare.domain;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.sql.Timestamp;
+
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+@Entity
+@Table(name = "trade_reviews")
+public class TradeReview {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "review_id")
+    private Long reviewId;
+
+    @ManyToOne
+    @JoinColumn(name = "food_id", referencedColumnName = "food_id")
+    private Food food;
+
+    @ManyToOne
+    @JoinColumn(name = "receiver_id", referencedColumnName = "user_id")
+    private User receiver;
+
+    @ManyToOne
+    @JoinColumn(name = "giver_id", referencedColumnName = "user_id")
+    private User giver;
+
+    @Column(name = "rating")
+    private Integer rating;
+
+    @Column(name = "comment")
+    private String comment;
+
+    @Column(name = "created_at")
+    private Timestamp createdAt;
+
+    @Column(name = "updated_at")
+    private Timestamp updatedAt;
+
+    // Getters and Setters
+}
